task-agent/requirements.txt
langchain>=1.0.0
langgraph>=1.0.0
langchain-deepseek>=1.0.1
# langchain-openai>=1.1.12
# langchain-anthropic>=1.4.0
pydantic>=2.12.5
python-dotenv>=1.0.1
